requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
cryptography==41.0.7
marshmallow==3.20.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv
pytest==7.4.4
